+++ packages/Angular/Generic/export-service/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @memberjunction/ng-export-service
 
+## 5.26.0
+
+### Patch Changes
+
+- @memberjunction/export-engine@5.26.0
+
+## 5.25.0
+
+### Patch Changes
+
+- @memberjunction/export-engine@5.25.0
+
 ## 5.24.0
 
 ### Patch Changes